reorganize tests
* move ref files to separate refernce folder
* stop Qasm tests running multiple times
