+++ test/terra/backends/qasm_simulator/qasm_algorithms.py
@@ -4,18 +4,16 @@
 #
 # This source code is licensed under the Apache License, Version 2.0 found in
 # the LICENSE.txt file in the root directory of this source tree.
-
 """
 QasmSimulator Integration Tests
 """
 
-from test.terra.utils import common
-from test.terra.utils import ref_algorithms
+from test.terra.reference import ref_algorithms
 from qiskit import compile
 from qiskit.providers.aer import QasmSimulator
 
 
-class QasmAlgorithmTests(common.QiskitAerTestCase):
+class QasmAlgorithmTests:
     """QasmSimulator algorithm tests in the default basis"""
 
     SIMULATOR = QasmSimulator()
@@ -27,8 +25,8 @@ class QasmAlgorithmTests(common.QiskitAerTestCase):
     def test_grovers_default_basis_gates(self):
         """Test grovers circuits compiling to backend default basis_gates."""
         shots = 2000
-        circuits = ref_algorithms.grovers_circuit(final_measure=True,
-                                                  allow_sampling=True)
+        circuits = ref_algorithms.grovers_circuit(
+            final_measure=True, allow_sampling=True)
         targets = ref_algorithms.grovers_counts(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
         result = self.SIMULATOR.run(qobj).result()
@@ -46,7 +44,7 @@ class QasmAlgorithmTests(common.QiskitAerTestCase):
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
 
-class QasmAlgorithmTestsWaltzBasis(common.QiskitAerTestCase):
+class QasmAlgorithmTestsWaltzBasis:
     """QasmSimulator algorithm tests in the Waltz u1,u2,u3,cx basis"""
 
     SIMULATOR = QasmSimulator()
@@ -58,10 +56,14 @@ class QasmAlgorithmTestsWaltzBasis(common.QiskitAerTestCase):
     def test_grovers_waltz_basis_gates(self):
         """Test grovers gate circuits compiling to u1,u2,u3,cx"""
         shots = 2000
-        circuits = ref_algorithms.grovers_circuit(final_measure=True,
-                                                  allow_sampling=True)
+        circuits = ref_algorithms.grovers_circuit(
+            final_measure=True, allow_sampling=True)
         targets = ref_algorithms.grovers_counts(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
         result = self.SIMULATOR.run(qobj).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -71,13 +73,17 @@ class QasmAlgorithmTestsWaltzBasis(common.QiskitAerTestCase):
         shots = 2000
         circuits = ref_algorithms.teleport_circuit()
         targets = ref_algorithms.teleport_counts(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
         result = self.SIMULATOR.run(qobj).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
 
-class QasmAlgorithmTestsMinimalBasis(common.QiskitAerTestCase):
+class QasmAlgorithmTestsMinimalBasis:
     """QasmSimulator algorithm tests in the minimal U,CX basis"""
 
     SIMULATOR = QasmSimulator()
@@ -89,10 +95,11 @@ class QasmAlgorithmTestsMinimalBasis(common.QiskitAerTestCase):
     def test_grovers_minimal_basis_gates(self):
         """Test grovers circuits compiling to u3,cx"""
         shots = 2000
-        circuits = ref_algorithms.grovers_circuit(final_measure=True,
-                                                  allow_sampling=True)
+        circuits = ref_algorithms.grovers_circuit(
+            final_measure=True, allow_sampling=True)
         targets = ref_algorithms.grovers_counts(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
         result = self.SIMULATOR.run(qobj).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -102,7 +109,8 @@ class QasmAlgorithmTestsMinimalBasis(common.QiskitAerTestCase):
         shots = 2000
         circuits = ref_algorithms.teleport_circuit()
         targets = ref_algorithms.teleport_counts(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
         result = self.SIMULATOR.run(qobj).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
