+++ test/terra/backends/qasm_simulator/qasm_cliffords.py
@@ -4,19 +4,17 @@
 #
 # This source code is licensed under the Apache License, Version 2.0 found in
 # the LICENSE.txt file in the root directory of this source tree.
-
 """
 QasmSimulator Integration Tests
 """
 
-from test.terra.utils import common
-from test.terra.utils import ref_1q_clifford
-from test.terra.utils import ref_2q_clifford
+from test.terra.reference import ref_1q_clifford
+from test.terra.reference import ref_2q_clifford
 from qiskit import compile
 from qiskit.providers.aer import QasmSimulator
 
 
-class QasmCliffordTests(common.QiskitAerTestCase):
+class QasmCliffordTests:
     """QasmSimulator Clifford gate tests in default basis."""
 
     SIMULATOR = QasmSimulator()
@@ -28,20 +26,24 @@ class QasmCliffordTests(common.QiskitAerTestCase):
     def test_h_gate_deterministic_default_basis_gates(self):
         """Test h-gate circuits compiling to backend default basis_gates."""
         shots = 100
-        circuits = ref_1q_clifford.h_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_1q_clifford.h_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_1q_clifford.h_gate_counts_deterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
     def test_h_gate_nondeterministic_default_basis_gates(self):
         """Test h-gate circuits compiling to backend default basis_gates."""
         shots = 2000
-        circuits = ref_1q_clifford.h_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_1q_clifford.h_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_1q_clifford.h_gate_counts_nondeterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
@@ -51,10 +53,12 @@ class QasmCliffordTests(common.QiskitAerTestCase):
     def test_x_gate_deterministic_default_basis_gates(self):
         """Test x-gate circuits compiling to backend default basis_gates."""
         shots = 100
-        circuits = ref_1q_clifford.x_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_1q_clifford.x_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_1q_clifford.x_gate_counts_deterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
@@ -64,10 +68,12 @@ class QasmCliffordTests(common.QiskitAerTestCase):
     def test_z_gate_deterministic_default_basis_gates(self):
         """Test z-gate circuits compiling to backend default basis_gates."""
         shots = 100
-        circuits = ref_1q_clifford.z_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_1q_clifford.z_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_1q_clifford.z_gate_counts_deterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
@@ -77,10 +83,12 @@ class QasmCliffordTests(common.QiskitAerTestCase):
     def test_y_gate_deterministic_default_basis_gates(self):
         """Test y-gate circuits compiling to backend default basis_gates."""
         shots = 100
-        circuits = ref_1q_clifford.y_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_1q_clifford.y_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_1q_clifford.y_gate_counts_deterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
@@ -90,20 +98,24 @@ class QasmCliffordTests(common.QiskitAerTestCase):
     def test_s_gate_deterministic_default_basis_gates(self):
         """Test s-gate circuits compiling to backend default basis_gates."""
         shots = 100
-        circuits = ref_1q_clifford.s_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_1q_clifford.s_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_1q_clifford.s_gate_counts_deterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
     def test_s_gate_nondeterministic_default_basis_gates(self):
         """Test s-gate circuits compiling to backend default basis_gates."""
         shots = 2000
-        circuits = ref_1q_clifford.s_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_1q_clifford.s_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_1q_clifford.s_gate_counts_nondeterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
@@ -113,20 +125,24 @@ class QasmCliffordTests(common.QiskitAerTestCase):
     def test_sdg_gate_deterministic_default_basis_gates(self):
         """Test sdg-gate circuits compiling to backend default basis_gates."""
         shots = 100
-        circuits = ref_1q_clifford.sdg_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_1q_clifford.sdg_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_1q_clifford.sdg_gate_counts_deterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
     def test_sdg_gate_nondeterministic_default_basis_gates(self):
         shots = 2000
         """Test sdg-gate circuits compiling to backend default basis_gates."""
-        circuits = ref_1q_clifford.sdg_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_1q_clifford.sdg_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_1q_clifford.sdg_gate_counts_nondeterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
@@ -136,20 +152,24 @@ class QasmCliffordTests(common.QiskitAerTestCase):
     def test_cx_gate_deterministic_default_basis_gates(self):
         """Test cx-gate circuits compiling to backend default basis_gates."""
         shots = 100
-        circuits = ref_2q_clifford.cx_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_2q_clifford.cx_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_2q_clifford.cx_gate_counts_deterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
     def test_cx_gate_nondeterministic_default_basis_gates(self):
         """Test cx-gate circuits compiling to backend default basis_gates."""
         shots = 2000
-        circuits = ref_2q_clifford.cx_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_2q_clifford.cx_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_2q_clifford.cx_gate_counts_nondeterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
@@ -159,20 +179,24 @@ class QasmCliffordTests(common.QiskitAerTestCase):
     def test_cz_gate_deterministic_default_basis_gates(self):
         """Test cz-gate circuits compiling to backend default basis_gates."""
         shots = 100
-        circuits = ref_2q_clifford.cz_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_2q_clifford.cz_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_2q_clifford.cz_gate_counts_deterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
     def test_cz_gate_nondeterministic_default_basis_gates(self):
         """Test cz-gate circuits compiling to backend default basis_gates."""
         shots = 2000
-        circuits = ref_2q_clifford.cz_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_2q_clifford.cz_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_2q_clifford.cz_gate_counts_nondeterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
@@ -182,25 +206,29 @@ class QasmCliffordTests(common.QiskitAerTestCase):
     def test_swap_gate_deterministic_default_basis_gates(self):
         """Test swap-gate circuits compiling to backend default basis_gates."""
         shots = 100
-        circuits = ref_2q_clifford.swap_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_2q_clifford.swap_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_2q_clifford.swap_gate_counts_deterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
     def test_swap_gate_nondeterministic_default_basis_gates(self):
         """Test swap-gate circuits compiling to backend default basis_gates."""
         shots = 2000
-        circuits = ref_2q_clifford.swap_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_2q_clifford.swap_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_2q_clifford.swap_gate_counts_nondeterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
 
-class QasmCliffordTestsWaltzBasis(common.QiskitAerTestCase):
+class QasmCliffordTestsWaltzBasis:
     """QasmSimulator Clifford gate tests in Waltz u1,u2,u3,cx basis."""
 
     SIMULATOR = QasmSimulator()
@@ -212,20 +240,32 @@ class QasmCliffordTestsWaltzBasis(common.QiskitAerTestCase):
     def test_h_gate_deterministic_waltz_basis_gates(self):
         """Test h-gate gate circuits compiling to u1,u2,u3,cx"""
         shots = 100
-        circuits = ref_1q_clifford.h_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_1q_clifford.h_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_1q_clifford.h_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
     def test_h_gate_nondeterministic_waltz_basis_gates(self):
         """Test h-gate gate circuits compiling to u1,u2,u3,cx"""
         shots = 2000
-        circuits = ref_1q_clifford.h_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_1q_clifford.h_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_1q_clifford.h_gate_counts_nondeterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
@@ -235,10 +275,16 @@ class QasmCliffordTestsWaltzBasis(common.QiskitAerTestCase):
     def test_x_gate_deterministic_waltz_basis_gates(self):
         """Test x-gate gate circuits compiling to u1,u2,u3,cx"""
         shots = 100
-        circuits = ref_1q_clifford.x_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_1q_clifford.x_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_1q_clifford.x_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
@@ -249,20 +295,29 @@ class QasmCliffordTestsWaltzBasis(common.QiskitAerTestCase):
     def test_z_gate_deterministic_waltz_basis_gates(self):
         """Test z-gate gate circuits compiling to u1,u2,u3,cx"""
         shots = 100
-        circuits = ref_1q_clifford.z_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_1q_clifford.z_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_1q_clifford.z_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
     def test_z_gate_deterministic_minimal_basis_gates(self):
         """Test z-gate gate circuits compiling to u3,cx"""
         shots = 100
-        circuits = ref_1q_clifford.z_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_1q_clifford.z_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_1q_clifford.z_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
@@ -272,20 +327,28 @@ class QasmCliffordTestsWaltzBasis(common.QiskitAerTestCase):
     def test_y_gate_deterministic_default_basis_gates(self):
         """Test y-gate circuits compiling to backend default basis_gates."""
         shots = 100
-        circuits = ref_1q_clifford.y_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_1q_clifford.y_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_1q_clifford.y_gate_counts_deterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
     def test_y_gate_deterministic_waltz_basis_gates(self):
         shots = 100
         """Test y-gate gate circuits compiling to u1,u2,u3,cx"""
-        circuits = ref_1q_clifford.y_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_1q_clifford.y_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_1q_clifford.y_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
@@ -295,20 +358,32 @@ class QasmCliffordTestsWaltzBasis(common.QiskitAerTestCase):
     def test_s_gate_deterministic_waltz_basis_gates(self):
         """Test s-gate gate circuits compiling to u1,u2,u3,cx"""
         shots = 100
-        circuits = ref_1q_clifford.s_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_1q_clifford.s_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_1q_clifford.s_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
     def test_s_gate_nondeterministic_waltz_basis_gates(self):
         """Test s-gate gate circuits compiling to u1,u2,u3,cx"""
         shots = 2000
-        circuits = ref_1q_clifford.s_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_1q_clifford.s_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_1q_clifford.s_gate_counts_nondeterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
@@ -318,20 +393,32 @@ class QasmCliffordTestsWaltzBasis(common.QiskitAerTestCase):
     def test_sdg_gate_deterministic_waltz_basis_gates(self):
         """Test sdg-gate gate circuits compiling to u1,u2,u3,cx"""
         shots = 100
-        circuits = ref_1q_clifford.sdg_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_1q_clifford.sdg_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_1q_clifford.sdg_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
     def test_sdg_gate_nondeterministic_waltz_basis_gates(self):
         """Test sdg-gate gate circuits compiling to u1,u2,u3,cx"""
         shots = 2000
-        circuits = ref_1q_clifford.sdg_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_1q_clifford.sdg_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_1q_clifford.sdg_gate_counts_nondeterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
@@ -341,20 +428,32 @@ class QasmCliffordTestsWaltzBasis(common.QiskitAerTestCase):
     def test_cx_gate_deterministic_waltz_basis_gates(self):
         shots = 100
         """Test cx-gate gate circuits compiling to u1,u2,u3,cx"""
-        circuits = ref_2q_clifford.cx_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_2q_clifford.cx_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_2q_clifford.cx_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
     def test_cx_gate_nondeterministic_waltz_basis_gates(self):
         """Test cx-gate gate circuits compiling to u1,u2,u3,cx"""
         shots = 2000
-        circuits = ref_2q_clifford.cx_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_2q_clifford.cx_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_2q_clifford.cx_gate_counts_nondeterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
@@ -364,20 +463,32 @@ class QasmCliffordTestsWaltzBasis(common.QiskitAerTestCase):
     def test_cz_gate_deterministic_waltz_basis_gates(self):
         """Test cz-gate gate circuits compiling to u1,u2,u3,cx"""
         shots = 100
-        circuits = ref_2q_clifford.cz_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_2q_clifford.cz_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_2q_clifford.cz_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
     def test_cz_gate_nondeterministic_waltz_basis_gates(self):
         """Test cz-gate gate circuits compiling to u1,u2,u3,cx"""
         shots = 2000
-        circuits = ref_2q_clifford.cz_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_2q_clifford.cz_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_2q_clifford.cz_gate_counts_nondeterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
@@ -387,25 +498,37 @@ class QasmCliffordTestsWaltzBasis(common.QiskitAerTestCase):
     def test_swap_gate_deterministic_waltz_basis_gates(self):
         """Test swap-gate gate circuits compiling to u1,u2,u3,cx"""
         shots = 100
-        circuits = ref_2q_clifford.swap_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_2q_clifford.swap_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_2q_clifford.swap_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
     def test_swap_gate_nondeterministic_waltz_basis_gates(self):
         """Test swap-gate gate circuits compiling to u1,u2,u3,cx"""
         shots = 2000
-        circuits = ref_2q_clifford.swap_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_2q_clifford.swap_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_2q_clifford.swap_gate_counts_nondeterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
 
-class QasmCliffordTestsMinimalBasis(common.QiskitAerTestCase):
+class QasmCliffordTestsMinimalBasis:
     """QasmSimulator Clifford gate tests in minimam U,CX basis."""
 
     SIMULATOR = QasmSimulator()
@@ -418,20 +541,26 @@ class QasmCliffordTestsMinimalBasis(common.QiskitAerTestCase):
     def test_h_gate_deterministic_minimal_basis_gates(self):
         """Test h-gate gate circuits compiling to u3,cx"""
         shots = 100
-        circuits = ref_1q_clifford.h_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_1q_clifford.h_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_1q_clifford.h_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
     def test_h_gate_nondeterministic_minimal_basis_gates(self):
         """Test h-gate gate circuits compiling to u3,cx"""
         shots = 2000
-        circuits = ref_1q_clifford.h_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_1q_clifford.h_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_1q_clifford.h_gate_counts_nondeterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
@@ -441,10 +570,13 @@ class QasmCliffordTestsMinimalBasis(common.QiskitAerTestCase):
     def test_x_gate_deterministic_minimal_basis_gates(self):
         """Test x-gate gate circuits compiling to u3,cx"""
         shots = 100
-        circuits = ref_1q_clifford.x_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_1q_clifford.x_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_1q_clifford.x_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
@@ -455,10 +587,13 @@ class QasmCliffordTestsMinimalBasis(common.QiskitAerTestCase):
     def test_z_gate_deterministic_minimal_basis_gates(self):
         """Test z-gate gate circuits compiling to u3,cx"""
         shots = 100
-        circuits = ref_1q_clifford.z_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_1q_clifford.z_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_1q_clifford.z_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
@@ -469,10 +604,13 @@ class QasmCliffordTestsMinimalBasis(common.QiskitAerTestCase):
     def test_y_gate_deterministic_minimal_basis_gates(self):
         """Test y-gate gate circuits compiling to u3,cx"""
         shots = 100
-        circuits = ref_1q_clifford.y_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_1q_clifford.y_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_1q_clifford.y_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
@@ -483,20 +621,26 @@ class QasmCliffordTestsMinimalBasis(common.QiskitAerTestCase):
     def test_s_gate_deterministic_minimal_basis_gates(self):
         """Test s-gate gate circuits compiling to u3,cx"""
         shots = 100
-        circuits = ref_1q_clifford.s_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_1q_clifford.s_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_1q_clifford.s_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
     def test_s_gate_nondeterministic_minimal_basis_gates(self):
         """Test s-gate gate circuits compiling to u3,cx"""
         shots = 2000
-        circuits = ref_1q_clifford.s_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_1q_clifford.s_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_1q_clifford.s_gate_counts_nondeterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
@@ -506,20 +650,26 @@ class QasmCliffordTestsMinimalBasis(common.QiskitAerTestCase):
     def test_sdg_gate_deterministic_minimal_basis_gates(self):
         """Test sdg-gate gate circuits compiling to u3,cx"""
         shots = 100
-        circuits = ref_1q_clifford.sdg_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_1q_clifford.sdg_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_1q_clifford.sdg_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
     def test_sdg_gate_nondeterministic_minimal_basis_gates(self):
         """Test sdg-gate gate circuits compiling to u3,cx"""
         shots = 2000
-        circuits = ref_1q_clifford.sdg_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_1q_clifford.sdg_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_1q_clifford.sdg_gate_counts_nondeterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
@@ -529,20 +679,26 @@ class QasmCliffordTestsMinimalBasis(common.QiskitAerTestCase):
     def test_cx_gate_deterministic_minimal_basis_gates(self):
         """Test cx-gate gate circuits compiling to u3,cx"""
         shots = 100
-        circuits = ref_2q_clifford.cx_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_2q_clifford.cx_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_2q_clifford.cx_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
     def test_cx_gate_nondeterministic_minimal_basis_gates(self):
         """Test cx-gate gate circuits compiling to u3,cx"""
         shots = 2000
-        circuits = ref_2q_clifford.cx_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_2q_clifford.cx_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_2q_clifford.cx_gate_counts_nondeterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
@@ -552,20 +708,26 @@ class QasmCliffordTestsMinimalBasis(common.QiskitAerTestCase):
     def test_cz_gate_deterministic_minimal_basis_gates(self):
         """Test cz-gate gate circuits compiling to u3,cx"""
         shots = 100
-        circuits = ref_2q_clifford.cz_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_2q_clifford.cz_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_2q_clifford.cz_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
     def test_cz_gate_nondeterministic_minimal_basis_gates(self):
         """Test cz-gate gate circuits compiling to u3,cx"""
         shots = 2000
-        circuits = ref_2q_clifford.cz_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_2q_clifford.cz_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_2q_clifford.cz_gate_counts_nondeterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
@@ -575,19 +737,25 @@ class QasmCliffordTestsMinimalBasis(common.QiskitAerTestCase):
     def test_swap_gate_deterministic_minimal_basis_gates(self):
         """Test swap-gate gate circuits compiling to u3,cx"""
         shots = 100
-        circuits = ref_2q_clifford.swap_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_2q_clifford.swap_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_2q_clifford.swap_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
     def test_swap_gate_nondeterministic_minimal_basis_gates(self):
         """Test swap-gate gate circuits compiling to u3,cx"""
         shots = 2000
-        circuits = ref_2q_clifford.swap_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_2q_clifford.swap_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_2q_clifford.swap_gate_counts_nondeterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
