+++ test/terra/backends/qasm_simulator/qasm_initialize.py
@@ -4,18 +4,16 @@
 #
 # This source code is licensed under the Apache License, Version 2.0 found in
 # the LICENSE.txt file in the root directory of this source tree.
-
 """
 QasmSimulator Integration Tests
 """
 
-from test.terra.utils import common
-from test.terra.utils import ref_initialize
+from test.terra.reference import ref_initialize
 from qiskit import compile
 from qiskit.providers.aer import QasmSimulator
 
 
-class QasmInitializeTests(common.QiskitAerTestCase):
+class QasmInitializeTests:
     """QasmSimulator initialize tests."""
 
     SIMULATOR = QasmSimulator()
@@ -32,7 +30,8 @@ class QasmInitializeTests(common.QiskitAerTestCase):
         circuits = ref_initialize.initialize_circuits_1(final_measure=True)
         targets = ref_initialize.initialize_counts_1(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
@@ -44,7 +43,8 @@ class QasmInitializeTests(common.QiskitAerTestCase):
         circuits = ref_initialize.initialize_circuits_2(final_measure=True)
         targets = ref_initialize.initialize_counts_2(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
@@ -54,6 +54,7 @@ class QasmInitializeTests(common.QiskitAerTestCase):
         circuits = ref_initialize.initialize_sampling_optimization()
         targets = ref_initialize.initialize_counts_sampling_optimization(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
