+++ test/terra/backends/qasm_simulator/qasm_method.py
@@ -4,14 +4,12 @@
 #
 # This source code is licensed under the Apache License, Version 2.0 found in
 # the LICENSE.txt file in the root directory of this source tree.
-
 """
 QasmSimulator Integration Tests
 """
 
-from test.terra.utils import common
-from test.terra.utils import ref_2q_clifford
-from test.terra.utils import ref_non_clifford
+from test.terra.reference import ref_2q_clifford
+from test.terra.reference import ref_non_clifford
 from qiskit import compile
 from qiskit.providers.aer import QasmSimulator
 from qiskit.providers.aer import AerError
@@ -21,7 +19,7 @@ from qiskit.providers.aer.noise.errors import pauli_error
 from qiskit.providers.aer.noise.errors import amplitude_damping_error
 
 
-class QasmMethodTests(common.QiskitAerTestCase):
+class QasmMethodTests:
     """QasmSimulator method option tests."""
 
     SIMULATOR = QasmSimulator()
@@ -34,48 +32,63 @@ class QasmMethodTests(common.QiskitAerTestCase):
         """Test statevector method is used for Clifford circuit"""
         # Test circuits
         shots = 100
-        circuits = ref_2q_clifford.cz_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_2q_clifford.cz_gate_circuits_deterministic(
+            final_measure=True)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
 
         def get_result():
-            return self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+            return self.SIMULATOR.run(
+                qobj, backend_options=self.BACKEND_OPTS).result()
 
         # Check simulation method
         method = self.BACKEND_OPTS.get('method')
         result = get_result()
         self.is_completed(result)
         if method == 'statevector':
-            self.compare_result_metadata(result, circuits, 'method', 'statevector')
+            self.compare_result_metadata(result, circuits, 'method',
+                                         'statevector')
         else:
-            self.compare_result_metadata(result, circuits, 'method', 'stabilizer')
+            self.compare_result_metadata(result, circuits, 'method',
+                                         'stabilizer')
 
     def test_backend_method_clifford_circuits_and_reset_noise(self):
         """Test statevector method is used for Clifford circuit"""
         # Test noise model
-        noise_circs = [[{"name": "reset", "qubits": [0]}],
-                       [{"name": "id", "qubits": [0]}]]
+        noise_circs = [[{
+            "name": "reset",
+            "qubits": [0]
+        }], [{
+            "name": "id",
+            "qubits": [0]
+        }]]
         noise_probs = [0.5, 0.5]
         error = QuantumError(zip(noise_circs, noise_probs))
         noise_model = NoiseModel()
-        noise_model.add_all_qubit_quantum_error(error, ['id', 'x', 'y', 'z', 'h', 's', 'sdg'])
+        noise_model.add_all_qubit_quantum_error(
+            error, ['id', 'x', 'y', 'z', 'h', 's', 'sdg'])
 
         # Test circuits
         shots = 100
-        circuits = ref_2q_clifford.cz_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_2q_clifford.cz_gate_circuits_deterministic(
+            final_measure=True)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
 
         def get_result():
-            return self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS,
-                                      noise_model=noise_model).result()
+            return self.SIMULATOR.run(
+                qobj,
+                backend_options=self.BACKEND_OPTS,
+                noise_model=noise_model).result()
 
         # Check simulation method
         method = self.BACKEND_OPTS.get('method')
         result = get_result()
         self.is_completed(result)
         if method == 'statevector':
-            self.compare_result_metadata(result, circuits, 'method', 'statevector')
+            self.compare_result_metadata(result, circuits, 'method',
+                                         'statevector')
         else:
-            self.compare_result_metadata(result, circuits, 'method', 'stabilizer')
+            self.compare_result_metadata(result, circuits, 'method',
+                                         'stabilizer')
 
     def test_backend_method_clifford_circuits_and_pauli_noise(self):
         """Test statevector method is used for Clifford circuit"""
@@ -86,21 +99,27 @@ class QasmMethodTests(common.QiskitAerTestCase):
 
         # Test circuits
         shots = 100
-        circuits = ref_2q_clifford.cz_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_2q_clifford.cz_gate_circuits_deterministic(
+            final_measure=True)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
 
         def get_result():
-            return self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS,
-                                      noise_model=noise_model).result()
+            return self.SIMULATOR.run(
+                qobj,
+                backend_options=self.BACKEND_OPTS,
+                noise_model=noise_model).result()
 
         # Check simulation method
         method = self.BACKEND_OPTS.get('method')
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         if method == 'statevector':
-            self.compare_result_metadata(result, circuits, 'method', 'statevector')
+            self.compare_result_metadata(result, circuits, 'method',
+                                         'statevector')
         else:
-            self.compare_result_metadata(result, circuits, 'method', 'stabilizer')
+            self.compare_result_metadata(result, circuits, 'method',
+                                         'stabilizer')
 
     def test_backend_method_clifford_circuits_and_unitary_noise(self):
         """Test statevector method is used for Clifford circuit"""
@@ -111,12 +130,15 @@ class QasmMethodTests(common.QiskitAerTestCase):
 
         # Test circuits
         shots = 100
-        circuits = ref_2q_clifford.cz_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_2q_clifford.cz_gate_circuits_deterministic(
+            final_measure=True)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
 
         def get_result():
-            return self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS,
-                                      noise_model=noise_model).result()
+            return self.SIMULATOR.run(
+                qobj,
+                backend_options=self.BACKEND_OPTS,
+                noise_model=noise_model).result()
 
         # Check simulation method
         method = self.BACKEND_OPTS.get('method')
@@ -125,23 +147,28 @@ class QasmMethodTests(common.QiskitAerTestCase):
         else:
             result = get_result()
             self.is_completed(result)
-            self.compare_result_metadata(result, circuits, 'method', 'statevector')
+            self.compare_result_metadata(result, circuits, 'method',
+                                         'statevector')
 
     def test_backend_method_clifford_circuits_and_kraus_noise(self):
         """Test statevector method is used for Clifford circuit"""
         # Noise Model
         error = amplitude_damping_error(0.5)
         noise_model = NoiseModel()
-        noise_model.add_all_qubit_quantum_error(error, ['id', 'x', 'y', 'z', 'h', 's', 'sdg'])
+        noise_model.add_all_qubit_quantum_error(
+            error, ['id', 'x', 'y', 'z', 'h', 's', 'sdg'])
 
         # Test circuits
         shots = 100
-        circuits = ref_2q_clifford.cz_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_2q_clifford.cz_gate_circuits_deterministic(
+            final_measure=True)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
 
         def get_result():
-            return self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS,
-                                      noise_model=noise_model).result()
+            return self.SIMULATOR.run(
+                qobj,
+                backend_options=self.BACKEND_OPTS,
+                noise_model=noise_model).result()
 
         # Check simulation method
         method = self.BACKEND_OPTS.get('method')
@@ -150,7 +177,8 @@ class QasmMethodTests(common.QiskitAerTestCase):
         else:
             result = get_result()
             self.is_completed(result)
-            self.compare_result_metadata(result, circuits, 'method', 'statevector')
+            self.compare_result_metadata(result, circuits, 'method',
+                                         'statevector')
 
     # ---------------------------------------------------------------------
     # Test non-Clifford circuits with clifford and non-clifford noise
@@ -159,11 +187,13 @@ class QasmMethodTests(common.QiskitAerTestCase):
         """Test statevector method is used for Clifford circuit"""
         # Test circuits
         shots = 100
-        circuits = ref_non_clifford.ccx_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_non_clifford.ccx_gate_circuits_deterministic(
+            final_measure=True)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
 
         def get_result():
-            return self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+            return self.SIMULATOR.run(
+                qobj, backend_options=self.BACKEND_OPTS).result()
 
         # Check simulation method
         method = self.BACKEND_OPTS.get('method')
@@ -172,26 +202,36 @@ class QasmMethodTests(common.QiskitAerTestCase):
         else:
             result = get_result()
             self.is_completed(result)
-            self.compare_result_metadata(result, circuits, 'method', 'statevector')
+            self.compare_result_metadata(result, circuits, 'method',
+                                         'statevector')
 
     def test_backend_method_nonclifford_circuit_and_reset_noise(self):
         """Test statevector method is used for Clifford circuit"""
         # Test noise model
-        noise_circs = [[{"name": "reset", "qubits": [0]}],
-                       [{"name": "id", "qubits": [0]}]]
+        noise_circs = [[{
+            "name": "reset",
+            "qubits": [0]
+        }], [{
+            "name": "id",
+            "qubits": [0]
+        }]]
         noise_probs = [0.5, 0.5]
         error = QuantumError(zip(noise_circs, noise_probs))
         noise_model = NoiseModel()
-        noise_model.add_all_qubit_quantum_error(error, ['id', 'x', 'y', 'z', 'h', 's', 'sdg'])
+        noise_model.add_all_qubit_quantum_error(
+            error, ['id', 'x', 'y', 'z', 'h', 's', 'sdg'])
 
         # Test circuits
         shots = 100
-        circuits = ref_non_clifford.ccx_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_non_clifford.ccx_gate_circuits_deterministic(
+            final_measure=True)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
 
         def get_result():
-            return self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS,
-                                      noise_model=noise_model).result()
+            return self.SIMULATOR.run(
+                qobj,
+                backend_options=self.BACKEND_OPTS,
+                noise_model=noise_model).result()
 
         # Check simulation method
         method = self.BACKEND_OPTS.get('method')
@@ -200,7 +240,8 @@ class QasmMethodTests(common.QiskitAerTestCase):
         else:
             result = get_result()
             self.is_completed(result)
-            self.compare_result_metadata(result, circuits, 'method', 'statevector')
+            self.compare_result_metadata(result, circuits, 'method',
+                                         'statevector')
 
     def test_backend_method_nonclifford_circuit_and_pauli_noise(self):
         """Test statevector method is used for Clifford circuit"""
@@ -211,12 +252,15 @@ class QasmMethodTests(common.QiskitAerTestCase):
 
         # Test circuits
         shots = 100
-        circuits = ref_non_clifford.ccx_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_non_clifford.ccx_gate_circuits_deterministic(
+            final_measure=True)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
 
         def get_result():
-            return self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS,
-                                      noise_model=noise_model).result()
+            return self.SIMULATOR.run(
+                qobj,
+                backend_options=self.BACKEND_OPTS,
+                noise_model=noise_model).result()
 
         # Check simulation method
         method = self.BACKEND_OPTS.get('method')
@@ -225,7 +269,8 @@ class QasmMethodTests(common.QiskitAerTestCase):
         else:
             result = get_result()
             self.is_completed(result)
-            self.compare_result_metadata(result, circuits, 'method', 'statevector')
+            self.compare_result_metadata(result, circuits, 'method',
+                                         'statevector')
 
     def test_backend_method_nonclifford_circuit_and_unitary_noise(self):
         """Test statevector method is used for Clifford circuit"""
@@ -236,12 +281,15 @@ class QasmMethodTests(common.QiskitAerTestCase):
 
         # Test circuits
         shots = 100
-        circuits = ref_non_clifford.ccx_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_non_clifford.ccx_gate_circuits_deterministic(
+            final_measure=True)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
 
         def get_result():
-            return self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS,
-                                      noise_model=noise_model).result()
+            return self.SIMULATOR.run(
+                qobj,
+                backend_options=self.BACKEND_OPTS,
+                noise_model=noise_model).result()
 
         # Check simulation method
         method = self.BACKEND_OPTS.get('method')
@@ -250,23 +298,28 @@ class QasmMethodTests(common.QiskitAerTestCase):
         else:
             result = get_result()
             self.is_completed(result)
-            self.compare_result_metadata(result, circuits, 'method', 'statevector')
+            self.compare_result_metadata(result, circuits, 'method',
+                                         'statevector')
 
     def test_backend_method_nonclifford_circuit_and_kraus_noise(self):
         """Test statevector method is used for Clifford circuit"""
         # Noise Model
         error = amplitude_damping_error(0.5)
         noise_model = NoiseModel()
-        noise_model.add_all_qubit_quantum_error(error, ['id', 'x', 'y', 'z', 'h', 's', 'sdg'])
+        noise_model.add_all_qubit_quantum_error(
+            error, ['id', 'x', 'y', 'z', 'h', 's', 'sdg'])
 
         # Test circuits
         shots = 100
-        circuits = ref_non_clifford.ccx_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_non_clifford.ccx_gate_circuits_deterministic(
+            final_measure=True)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
 
         def get_result():
-            return self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS,
-                                      noise_model=noise_model).result()
+            return self.SIMULATOR.run(
+                qobj,
+                backend_options=self.BACKEND_OPTS,
+                noise_model=noise_model).result()
 
         # Check simulation method
         method = self.BACKEND_OPTS.get('method')
@@ -275,4 +328,5 @@ class QasmMethodTests(common.QiskitAerTestCase):
         else:
             result = get_result()
             self.is_completed(result)
-            self.compare_result_metadata(result, circuits, 'method', 'statevector')
+            self.compare_result_metadata(result, circuits, 'method',
+                                         'statevector')
