+++ test/terra/backends/qasm_simulator/qasm_noncliffords.py
@@ -4,18 +4,16 @@
 #
 # This source code is licensed under the Apache License, Version 2.0 found in
 # the LICENSE.txt file in the root directory of this source tree.
-
 """
 QasmSimulator Integration Tests
 """
 
-from test.terra.utils import common
-from test.terra.utils import ref_non_clifford
+from test.terra.reference import ref_non_clifford
 from qiskit import compile
 from qiskit.providers.aer import QasmSimulator
 
 
-class QasmNonCliffordTests(common.QiskitAerTestCase):
+class QasmNonCliffordTests:
     """QasmSimulator non-Clifford gate tests in default basis."""
 
     SIMULATOR = QasmSimulator()
@@ -27,7 +25,8 @@ class QasmNonCliffordTests(common.QiskitAerTestCase):
     def test_t_gate_deterministic_default_basis_gates(self):
         """Test t-gate circuits compiling to backend default basis_gates."""
         shots = 100
-        circuits = ref_non_clifford.t_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_non_clifford.t_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_non_clifford.t_gate_counts_deterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
         result = self.SIMULATOR.run(qobj).result()
@@ -37,7 +36,8 @@ class QasmNonCliffordTests(common.QiskitAerTestCase):
     def test_t_gate_nondeterministic_default_basis_gates(self):
         """Test t-gate circuits compiling to backend default basis_gates."""
         shots = 2000
-        circuits = ref_non_clifford.t_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_non_clifford.t_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_non_clifford.t_gate_counts_nondeterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
         result = self.SIMULATOR.run(qobj).result()
@@ -50,7 +50,8 @@ class QasmNonCliffordTests(common.QiskitAerTestCase):
     def test_tdg_gate_deterministic_default_basis_gates(self):
         """Test tdg-gate circuits compiling to backend default basis_gates."""
         shots = 100
-        circuits = ref_non_clifford.tdg_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_non_clifford.tdg_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_non_clifford.tdg_gate_counts_deterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
         result = self.SIMULATOR.run(qobj).result()
@@ -60,7 +61,8 @@ class QasmNonCliffordTests(common.QiskitAerTestCase):
     def test_tdg_gate_nondeterministic_default_basis_gates(self):
         """Test tdg-gate circuits compiling to backend default basis_gates."""
         shots = 2000
-        circuits = ref_non_clifford.tdg_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_non_clifford.tdg_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_non_clifford.tdg_gate_counts_nondeterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
         result = self.SIMULATOR.run(qobj).result()
@@ -73,7 +75,8 @@ class QasmNonCliffordTests(common.QiskitAerTestCase):
     def test_ccx_gate_deterministic_default_basis_gates(self):
         """Test ccx-gate circuits compiling to backend default basis_gates."""
         shots = 100
-        circuits = ref_non_clifford.ccx_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_non_clifford.ccx_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_non_clifford.ccx_gate_counts_deterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
         result = self.SIMULATOR.run(qobj).result()
@@ -83,7 +86,8 @@ class QasmNonCliffordTests(common.QiskitAerTestCase):
     def test_ccx_gate_nondeterministic_default_basis_gates(self):
         """Test ccx-gate circuits compiling to backend default basis_gates."""
         shots = 2000
-        circuits = ref_non_clifford.ccx_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_non_clifford.ccx_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_non_clifford.ccx_gate_counts_nondeterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
         result = self.SIMULATOR.run(qobj).result()
@@ -91,7 +95,7 @@ class QasmNonCliffordTests(common.QiskitAerTestCase):
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
 
-class QasmNonCliffordTestsWaltzBasis(common.QiskitAerTestCase):
+class QasmNonCliffordTestsWaltzBasis:
     """QasmSimulator non-Clifford gate tests in minimal u1,u2,u3,cx basis."""
 
     SIMULATOR = QasmSimulator()
@@ -103,9 +107,14 @@ class QasmNonCliffordTestsWaltzBasis(common.QiskitAerTestCase):
     def test_t_gate_deterministic_waltz_basis_gates(self):
         """Test t-gate gate circuits compiling to u1,u2,u3,cx"""
         shots = 100
-        circuits = ref_non_clifford.t_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_non_clifford.t_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_non_clifford.t_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
         result = self.SIMULATOR.run(qobj).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
@@ -113,9 +122,14 @@ class QasmNonCliffordTestsWaltzBasis(common.QiskitAerTestCase):
     def test_t_gate_nondeterministic_waltz_basis_gates(self):
         """Test t-gate gate circuits compiling to u1,u2,u3,cx"""
         shots = 2000
-        circuits = ref_non_clifford.t_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_non_clifford.t_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_non_clifford.t_gate_counts_nondeterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
         result = self.SIMULATOR.run(qobj).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -126,9 +140,14 @@ class QasmNonCliffordTestsWaltzBasis(common.QiskitAerTestCase):
     def test_tdg_gate_deterministic_waltz_basis_gates(self):
         """Test tdg-gate gate circuits compiling to u1,u2,u3,cx"""
         shots = 100
-        circuits = ref_non_clifford.tdg_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_non_clifford.tdg_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_non_clifford.tdg_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
         result = self.SIMULATOR.run(qobj).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
@@ -136,9 +155,14 @@ class QasmNonCliffordTestsWaltzBasis(common.QiskitAerTestCase):
     def test_tdg_gate_nondeterministic_waltz_basis_gates(self):
         """Test tdg-gate gate circuits compiling to u1,u2,u3,cx"""
         shots = 2000
-        circuits = ref_non_clifford.tdg_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_non_clifford.tdg_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_non_clifford.tdg_gate_counts_nondeterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
         result = self.SIMULATOR.run(qobj).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -149,9 +173,14 @@ class QasmNonCliffordTestsWaltzBasis(common.QiskitAerTestCase):
     def test_ccx_gate_deterministic_waltz_basis_gates(self):
         """Test ccx-gate gate circuits compiling to u1,u2,u3,cx"""
         shots = 100
-        circuits = ref_non_clifford.ccx_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_non_clifford.ccx_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_non_clifford.ccx_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
         result = self.SIMULATOR.run(qobj).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
@@ -159,15 +188,20 @@ class QasmNonCliffordTestsWaltzBasis(common.QiskitAerTestCase):
     def test_ccx_gate_nondeterministic_waltz_basis_gates(self):
         """Test ccx-gate gate circuits compiling to u1,u2,u3,cx"""
         shots = 2000
-        circuits = ref_non_clifford.ccx_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_non_clifford.ccx_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_non_clifford.ccx_gate_counts_nondeterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u1', 'u2', 'u3', 'cx'])
+        qobj = compile(
+            circuits,
+            self.SIMULATOR,
+            shots=shots,
+            basis_gates=['u1', 'u2', 'u3', 'cx'])
         result = self.SIMULATOR.run(qobj).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
 
-class QasmNonCliffordTestsMinimalBasis(common.QiskitAerTestCase):
+class QasmNonCliffordTestsMinimalBasis:
     """QasmSimulator non-Clifford gate tests in minimal U,CX basis."""
 
     SIMULATOR = QasmSimulator()
@@ -179,9 +213,11 @@ class QasmNonCliffordTestsMinimalBasis(common.QiskitAerTestCase):
     def test_t_gate_deterministic_minimal_basis_gates(self):
         """Test t-gate gate circuits compiling to u3,cx"""
         shots = 100
-        circuits = ref_non_clifford.t_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_non_clifford.t_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_non_clifford.t_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
         result = self.SIMULATOR.run(qobj).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
@@ -189,9 +225,11 @@ class QasmNonCliffordTestsMinimalBasis(common.QiskitAerTestCase):
     def test_t_gate_nondeterministic_minimal_basis_gates(self):
         """Test t-gate gate circuits compiling to u3,cx"""
         shots = 2000
-        circuits = ref_non_clifford.t_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_non_clifford.t_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_non_clifford.t_gate_counts_nondeterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
         result = self.SIMULATOR.run(qobj).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -202,9 +240,11 @@ class QasmNonCliffordTestsMinimalBasis(common.QiskitAerTestCase):
     def test_tdg_gate_deterministic_minimal_basis_gates(self):
         """Test tdg-gate gate circuits compiling to u3,cx"""
         shots = 100
-        circuits = ref_non_clifford.tdg_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_non_clifford.tdg_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_non_clifford.tdg_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
         result = self.SIMULATOR.run(qobj).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
@@ -212,9 +252,11 @@ class QasmNonCliffordTestsMinimalBasis(common.QiskitAerTestCase):
     def test_tdg_gate_nondeterministic_minimal_basis_gates(self):
         """Test tdg-gate gate circuits compiling to u3,cx"""
         shots = 2000
-        circuits = ref_non_clifford.tdg_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_non_clifford.tdg_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_non_clifford.tdg_gate_counts_nondeterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
         result = self.SIMULATOR.run(qobj).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -225,9 +267,11 @@ class QasmNonCliffordTestsMinimalBasis(common.QiskitAerTestCase):
     def test_ccx_gate_deterministic_minimal_basis_gates(self):
         """Test ccx-gate gate circuits compiling to u3,cx"""
         shots = 100
-        circuits = ref_non_clifford.ccx_gate_circuits_deterministic(final_measure=True)
+        circuits = ref_non_clifford.ccx_gate_circuits_deterministic(
+            final_measure=True)
         targets = ref_non_clifford.ccx_gate_counts_deterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
         result = self.SIMULATOR.run(qobj).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
@@ -235,9 +279,11 @@ class QasmNonCliffordTestsMinimalBasis(common.QiskitAerTestCase):
     def test_ccx_gate_nondeterministic_minimal_basis_gates(self):
         """Test ccx-gate gate circuits compiling to u3,cx"""
         shots = 2000
-        circuits = ref_non_clifford.ccx_gate_circuits_nondeterministic(final_measure=True)
+        circuits = ref_non_clifford.ccx_gate_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_non_clifford.ccx_gate_counts_nondeterministic(shots)
-        qobj = compile(circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+        qobj = compile(
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
         result = self.SIMULATOR.run(qobj).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
