+++ test/terra/backends/qasm_simulator/qasm_reset.py
@@ -4,18 +4,16 @@
 #
 # This source code is licensed under the Apache License, Version 2.0 found in
 # the LICENSE.txt file in the root directory of this source tree.
-
 """
 QasmSimulator Integration Tests
 """
 
-from test.terra.utils import common
-from test.terra.utils import ref_reset
+from test.terra.reference import ref_reset
 from qiskit import compile
 from qiskit.providers.aer import QasmSimulator
 
 
-class QasmResetTests(common.QiskitAerTestCase):
+class QasmResetTests:
     """QasmSimulator reset tests."""
 
     SIMULATOR = QasmSimulator()
@@ -32,7 +30,8 @@ class QasmResetTests(common.QiskitAerTestCase):
         circuits = ref_reset.reset_circuits_deterministic(final_measure=True)
         targets = ref_reset.reset_counts_deterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0)
 
@@ -41,10 +40,12 @@ class QasmResetTests(common.QiskitAerTestCase):
         # For statevector output we can combine deterministic and non-deterministic
         # count output circuits
         shots = 2000
-        circuits = ref_reset.reset_circuits_nondeterministic(final_measure=True)
+        circuits = ref_reset.reset_circuits_nondeterministic(
+            final_measure=True)
         targets = ref_reset.reset_counts_nondeterministic(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
@@ -54,6 +55,7 @@ class QasmResetTests(common.QiskitAerTestCase):
         circuits = ref_reset.reset_sampling_optimization()
         targets = ref_reset.reset_counts_sampling_optimization(shots)
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        result = self.SIMULATOR.run(qobj, backend_options=self.BACKEND_OPTS).result()
+        result = self.SIMULATOR.run(
+            qobj, backend_options=self.BACKEND_OPTS).result()
         self.is_completed(result)
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
