+++ test/terra/backends/qasm_simulator/qasm_thread_management.py
@@ -11,7 +11,6 @@ QasmSimulator Integration Tests
 import multiprocessing
 import psutil
 
-from test.terra.utils import common
 from test.benchmark.tools import quantum_volume_circuit
 from qiskit import compile
 from qiskit.providers.aer import QasmSimulator
@@ -19,10 +18,11 @@ from qiskit.providers.aer.noise import NoiseModel
 from qiskit.providers.aer.noise.errors.standard_errors import pauli_error
 
 
-class QasmThreadManagementTests(common.QiskitAerTestCase):
+class QasmThreadManagementTests:
     """QasmSimulator thread tests."""
 
     SIMULATOR = QasmSimulator()
+    BACKEND_OPTS = {}
 
     def dummy_noise_model(self):
         """Return test reset noise model"""
@@ -37,7 +37,9 @@ class QasmThreadManagementTests(common.QiskitAerTestCase):
         shots = 100
         circuit = quantum_volume_circuit(4, 1, measure=True, seed=0)
         qobj = compile(circuit, self.SIMULATOR, shots=shots)
-        backend_opts = {}
+
+        backend_opts = self.BACKEND_OPTS.copy()
+
         result = self.SIMULATOR.run(
             qobj, backend_options=backend_opts).result()
         if result.metadata['omp_enabled']:
@@ -62,7 +64,9 @@ class QasmThreadManagementTests(common.QiskitAerTestCase):
         shots = 100
         circuit = quantum_volume_circuit(4, 1, measure=True, seed=0)
         qobj = compile(circuit, self.SIMULATOR, shots=shots)
-        backend_opts = {}
+
+        backend_opts = self.BACKEND_OPTS.copy()
+
         result = self.SIMULATOR.run(
             qobj, backend_options=backend_opts).result()
         system_memory = int(psutil.virtual_memory().total / 1024 / 1024)
@@ -81,7 +85,10 @@ class QasmThreadManagementTests(common.QiskitAerTestCase):
         shots = 100
         circuit = quantum_volume_circuit(4, 1, measure=True, seed=0)
         qobj = compile(circuit, self.SIMULATOR, shots=shots)
-        backend_opts = {'max_memory_mb': 128}
+
+        backend_opts = self.BACKEND_OPTS.copy()
+        backend_opts['max_memory_mb'] = 128
+
         result = self.SIMULATOR.run(
             qobj, backend_options=backend_opts).result()
         self.assertEqual(
@@ -98,10 +105,11 @@ class QasmThreadManagementTests(common.QiskitAerTestCase):
         for _ in range(experiments):
             circuits.append(quantum_volume_circuit(4, 1, measure=True, seed=0))
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        backend_opts = {
-            'max_parallel_experiments': experiments,
-            'max_memory_mb': 1024
-        }
+
+        backend_opts = self.BACKEND_OPTS.copy()
+        backend_opts['max_parallel_experiments'] = experiments
+        backend_opts['max_memory_mb'] = 1024
+
         result = self.SIMULATOR.run(
             qobj, backend_options=backend_opts).result()
         if result.metadata['omp_enabled']:
@@ -131,10 +139,11 @@ class QasmThreadManagementTests(common.QiskitAerTestCase):
                 quantum_volume_circuit(16, 1, measure=True,
                                        seed=0))  # 2 MB for each
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        backend_opts = {
-            'max_parallel_experiments': experiments,
-            'max_memory_mb': 1
-        }
+
+        backend_opts = self.BACKEND_OPTS.copy()
+        backend_opts['max_parallel_experiments'] = experiments
+        backend_opts['max_memory_mb'] = 1
+
         result = self.SIMULATOR.run(
             qobj, backend_options=backend_opts).result()
         if result.metadata['omp_enabled']:
@@ -166,10 +175,11 @@ class QasmThreadManagementTests(common.QiskitAerTestCase):
                 quantum_volume_circuit(4, 1, measure=True,
                                        seed=0))  # 2 MB for each
         qobj = compile(circuits, self.SIMULATOR, shots=shots)
-        backend_opts = {
-            'max_parallel_experiments': experiments,
-            'max_memory_mb': 1024
-        }
+
+        backend_opts = self.BACKEND_OPTS.copy()
+        backend_opts['max_parallel_experiments'] = experiments
+        backend_opts['max_memory_mb'] = 1024
+
         result = self.SIMULATOR.run(
             qobj, backend_options=backend_opts).result()
         if result.metadata['omp_enabled']:
@@ -189,10 +199,11 @@ class QasmThreadManagementTests(common.QiskitAerTestCase):
         shots = multiprocessing.cpu_count()
         circuit = quantum_volume_circuit(4, 1, measure=True, seed=0)
         qobj = compile(circuit, self.SIMULATOR, shots=shots)
-        backend_opts = {
-            'max_parallel_shots': shots,
-            'noise_model': self.dummy_noise_model()
-        }
+
+        backend_opts = self.BACKEND_OPTS.copy()
+        backend_opts['max_parallel_shots'] = shots
+        backend_opts['noise_model'] = self.dummy_noise_model()
+
         result = self.SIMULATOR.run(
             qobj, backend_options=backend_opts).result()
         if result.metadata['omp_enabled']:
@@ -217,7 +228,10 @@ class QasmThreadManagementTests(common.QiskitAerTestCase):
         shots = multiprocessing.cpu_count()
         circuit = quantum_volume_circuit(4, 1, measure=True, seed=0)
         qobj = compile(circuit, self.SIMULATOR, shots=shots)
-        backend_opts = {'max_parallel_shots': shots}
+
+        backend_opts = self.BACKEND_OPTS.copy()
+        backend_opts['max_parallel_shots'] = shots
+
         result = self.SIMULATOR.run(
             qobj, backend_options=backend_opts).result()
         if result.metadata['omp_enabled']:
@@ -244,10 +258,11 @@ class QasmThreadManagementTests(common.QiskitAerTestCase):
             return
         circuit = quantum_volume_circuit(4, 1, measure=True, seed=0)
         qobj = compile(circuit, self.SIMULATOR, shots=shots)
-        backend_opts = {
-            'max_parallel_shots': multiprocessing.cpu_count(),
-            'noise_model': self.dummy_noise_model()
-        }
+
+        backend_opts = self.BACKEND_OPTS.copy()
+        backend_opts['max_parallel_shots'] = multiprocessing.cpu_count()
+        backend_opts['noise_model'] = self.dummy_noise_model()
+
         result = self.SIMULATOR.run(
             qobj, backend_options=backend_opts).result()
         if result.metadata['omp_enabled']:
@@ -272,11 +287,12 @@ class QasmThreadManagementTests(common.QiskitAerTestCase):
         shots = multiprocessing.cpu_count()
         circuit = quantum_volume_circuit(16, 1, measure=True, seed=0)
         qobj = compile(circuit, self.SIMULATOR, shots=shots)
-        backend_opts = {
-            'max_parallel_shots': shots,
-            'noise_model': self.dummy_noise_model(),
-            'max_memory_mb': 1
-        }
+
+        backend_opts = self.BACKEND_OPTS.copy()
+        backend_opts['max_parallel_shots'] = shots
+        backend_opts['noise_model'] = self.dummy_noise_model()
+        backend_opts['max_memory_mb'] = 1
+
         result = self.SIMULATOR.run(
             qobj, backend_options=backend_opts).result()
         if result.metadata['omp_enabled']:
