+++ test/terra/backends/test_qasm_stabilizer_simulator.py
@@ -10,16 +10,18 @@ QasmSimulator Integration Tests
 """
 
 import unittest
-from test.terra.utils.qasm_simulator.qasm_method import QasmMethodTests
-from test.terra.utils.qasm_simulator.qasm_measure import QasmMeasureTests
-from test.terra.utils.qasm_simulator.qasm_reset import QasmResetTests
-from test.terra.utils.qasm_simulator.qasm_conditional import QasmConditionalTests
-from test.terra.utils.qasm_simulator.qasm_cliffords import QasmCliffordTests
-from test.terra.utils.qasm_simulator.qasm_algorithms import QasmAlgorithmTests
-from test.terra.utils.qasm_simulator.qasm_extra import QasmExtraTests
+from test.terra import common
+from test.terra.backends.qasm_simulator.qasm_method import QasmMethodTests
+from test.terra.backends.qasm_simulator.qasm_measure import QasmMeasureTests
+from test.terra.backends.qasm_simulator.qasm_reset import QasmResetTests
+from test.terra.backends.qasm_simulator.qasm_conditional import QasmConditionalTests
+from test.terra.backends.qasm_simulator.qasm_cliffords import QasmCliffordTests
+from test.terra.backends.qasm_simulator.qasm_algorithms import QasmAlgorithmTests
+from test.terra.backends.qasm_simulator.qasm_extra import QasmExtraTests
 
 
-class TestQasmStabilizerSimulator(QasmMethodTests,
+class TestQasmStabilizerSimulator(common.QiskitAerTestCase,
+                                  QasmMethodTests,
                                   QasmMeasureTests,
                                   QasmResetTests,
                                   QasmConditionalTests,
