+++ test/terra/noise/test_quantum_error.py
@@ -9,7 +9,7 @@ QuantumError class tests
 """
 
 import unittest
-from test.terra.utils import common
+from test.terra import common
 import numpy as np
 
 from qiskit.quantum_info.operators.channel.superop import SuperOp
