+++ test/terra/noise/test_readout_error.py
@@ -12,7 +12,7 @@ import unittest
 
 import numpy as np
 
-from test.terra.utils import common
+from test.terra import common
 from qiskit.providers.aer.noise.noiseerror import NoiseError
 from qiskit.providers.aer.noise.errors.readout_error import ReadoutError
