+++ test/terra/reference/ref_1q_clifford.py
